* commit '7cb7e30a9a2b2c1db007cade0251202008259fd7':
  CAS-10646 : Duplicate of CAS-10264 for the 5.1 branch

+++ code/synthesis/ImagerObjects/SynthesisDeconvolver.cc
@@ -502,7 +502,7 @@ namespace casa { //# NAMESPACE CASA - BEGIN
 
       // Get the number of mask pixels (sum) and send to the logger.
       Float masksum = itsImages->getMaskSum();
-      Int npix = (itsImages->getShape()).product();
+      Float npix = (itsImages->getShape()).product();
       os << "[" << itsImages->getName() << "] Number of pixels in the clean mask : " << masksum << " out of a total of " << npix << " pixels. [ " << 100.0 * masksum/npix << " % ]" << LogIO::POST;
 
       maskchanged=True;
